Renaming internal include directory to match external installation directory
