+++ CMakeLists.txt
@@ -101,20 +101,20 @@ if(GIT_FOUND)
 endif()
 
 # install headers
-install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/RTI.hpp" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/primitives.hpp" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/ray.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/MOABRay.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/MOABDirectAccess.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/Vec3da.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/constants.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/sys.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/Vec3fa.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/Vec3ba.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/Vec3.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/embree3.hpp" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/embree4.hpp" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/embree_interface.hpp" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double_down/RTI.hpp" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double_down/primitives.hpp" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double_down/ray.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double_down/MOABRay.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double_down/MOABDirectAccess.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double_down/Vec3da.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double_down/constants.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double_down/sys.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double_down/Vec3fa.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double_down/Vec3ba.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double_down/Vec3.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double_down/embree3.hpp" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double_down/embree4.hpp" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double_down/embree_interface.hpp" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
 
 # Tests
 enable_testing()
@@ -124,8 +124,8 @@ configure_file("${CMAKE_CURRENT_SOURCE_DIR}/cmake/ddConfig.cmake.in" "${CMAKE_CU
 install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/cmake/ddConfig.cmake" DESTINATION "${DD_CMAKE_CONFIG_LOCATION}")
 configure_file("${CMAKE_CURRENT_SOURCE_DIR}/cmake/ddConfigVersion.cmake.in" "${CMAKE_CURRENT_SOURCE_DIR}/cmake/ddConfigVersion.cmake")
 install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/cmake/ddConfigVersion.cmake" DESTINATION "${DD_CMAKE_CONFIG_LOCATION}")
-configure_file("${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/version.h.in" "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/version.h")
-install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/version.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+configure_file("${CMAKE_CURRENT_SOURCE_DIR}/include/double_down/version.h.in" "${CMAKE_CURRENT_SOURCE_DIR}/include/double_down/version.h")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double_down/version.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
 
 add_subdirectory(docs)
 add_subdirectory(tools)
