+++ include/double_down/primitives.hpp
@@ -14,7 +14,7 @@
 #include "MOABDirectAccess.h"
 
 // Embree
-#include "double-down/embree_interface.hpp"
+#include "double_down/embree_interface.hpp"
 namespace double_down {
 
 /*! Structure with triangle information used in Embree */
