+++ test/test_closest.cpp
@@ -4,7 +4,7 @@
 
 #include "test_utils.hpp"
 
-#include "double-down/RTI.hpp"
+#include "double_down/RTI.hpp"
 
 using namespace double_down;
