+++ test/test_mb.cpp
@@ -2,7 +2,7 @@
 
 #include <iostream>
 
-#include "double-down/RTI.hpp"
+#include "double_down/RTI.hpp"
 
 #include "test_utils.hpp"
